@@ -11,20 +11,26 @@ import { FooterComponent } from './components/footer/footer.component';
 import { HomeComponent } from './components/home/home.component';
 import { RouterModule  } from '@angular/router';
 import { ReactiveFormsModule } from '@angular/forms';
+
 import { FilterPipe } from './pipes/filter.pipe';
 import { ClientService } from './services/client/client.service';
-import { ProductService } from './services/Products/Product.service';
+import { ProductService } from './services/products/product.service';
 import { CreateComponent } from './components/sales/create/create.component';
 import { ListComponent } from './components/sales/list/list.component';
-import { SaleService } from './services/Sales/sale.service';
-//import { HttpClientModule }       from '@angular/common/http';
+import { SaleService } from './services/sales/sale.service';
+import { HttpClientModule } from '@angular/common/http';
+import { UsersService } from './services/users.service';
+import { LoginComponent } from './components/login/login.component';
+import { AuthenticationGuard} from './guards/authentication.guard';
+
 
 var roots = [
-{ path: 'sales', component: SalesComponent},
-{ path: 'clients', component: ClientsComponent},
-{ path: 'inventory', component: InventoryComponent},
-{ path: 'home', component: HomeComponent},
-{ path: '', redirectTo: '/home', pathMatch: 'full' }
+  { path: 'sales', component: SalesComponent, canActivate : [AuthenticationGuard] },
+  { path: 'clients', component: ClientsComponent, canActivate : [AuthenticationGuard] },
+  { path: 'inventory', component: InventoryComponent, canActivate : [AuthenticationGuard] },
+  { path: 'home', component: HomeComponent },
+  { path: 'login', component: LoginComponent},
+  { path: '',   redirectTo: '/home', pathMatch: 'full' }
 ];
 
 @NgModule({
@@ -38,16 +44,18 @@ var roots = [
     HomeComponent,
     FilterPipe,
     CreateComponent,
-    ListComponent
+    ListComponent,
+    LoginComponent
   ],
+
   imports: [
     BrowserModule,
     FormsModule,
     RouterModule.forRoot(roots),
-    //HttpClientModule
+    HttpClientModule
   ],
 
-  providers: [ClientService, ProductService,SaleService],
+  providers: [ClientService, ProductService,SaleService, AuthenticationGuard, UsersService],
   bootstrap: [AppComponent],
 })
 export class AppModule { }
