@@ -13,33 +13,47 @@ export class ClientsComponent implements OnInit {
   user: {name: "", lastName: "", email: ""};
   action: string = "default";
 
-  onSave(user){
-  if (this.action == "edit"){
-
-    }
-    if (this.action == "create"){
-      this.users.push(user);
-    }
-  }
-
   onCreate(){
     this.user = {name: "", lastName: "", email: ""};
     this.action = "create";
-  }
-  onDelete(index){
-    this.users.splice(index, 1);
   }
 
   onEdit(user){
     this.user = user;
     this.action = "edit";
   }
-  constructor(private clientService: ClientService) {
+  constructor(private clientService: ClientService) {}
+
+  ngOnInit() {
+    console.log("Hola Clientes");
+    this.onFind();
 
   }
 
-  ngOnInit() {
-    this.users = this.clientService.clients;
+  onFind(){
+    this.clientService.find().subscribe((res:any) => {
+      this.users = res.body;
+      console.log(res.body);
+   });
+  }
+
+  onSave(user){
+    if (this.action == "edit"){
+      this.clientService.updateOne(user).subscribe((res:any) => {
+        this.onFind();
+      });
+    }
+    if (this.action == "create"){
+      this.clientService.insertOne(user).subscribe((res:any) => {
+        this.onFind();
+      });
+    }
+  }
+
+  onDelete(id){
+    this.clientService.deleteOne(id).subscribe((res:any) => {
+      this.onFind();
+    });
   }
 
 
