@@ -1,14 +1,17 @@
 import { Component, OnInit } from '@angular/core';
 
+
 @Component({
   selector: 'app-sales',
   templateUrl: './sales.component.html',
   styleUrls: ['./sales.component.css']
 })
-export class SalesComponent implements OnInit {
-  constructor() { }
 
+export class SalesComponent implements OnInit {
+
+  constructor() { }
   ngOnInit() {
+    console.log("Hola Pedidos");
   }
 
 }
