@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class ClientService {
@@ -9,6 +10,24 @@ export class ClientService {
   {name: "Juan Manuel ", lastName: "Medrano", email: "Juan@email.com"},
   ];
 
-  constructor() { }
+  constructor(private http:HttpClient) { }
+
+  find(){
+    console.log("Hola clientes");
+    return this.http.get('http://localhost:3000/api/clients');
+  }
+
+  insertOne(obj){
+    return this.http.post('http://localhost:3000/api/clients', obj);
+  }
+
+  updateOne(obj){
+    return this.http.put('http://localhost:3000/api/clients', obj);
+  }
+
+  deleteOne(id){
+    return this.http.delete('http://localhost:3000/api/clients/' + id);
+  }
+
 
 }
