@@ -1,4 +1,5 @@
 import { Injectable } from '@angular/core';
+import { HttpClient } from '@angular/common/http';
 
 @Injectable()
 export class ProductService {
@@ -9,7 +10,24 @@ export class ProductService {
   {id:4, name: "Cerveza", brand: "Lupulo", provider: "AGD", quantity: "100", price: "23"},
   {id:5, name: "Frijol", brand: "Mas", provider: "NUUE", quantity: "300", price: "10"},
   ];
-  
-  constructor() { }
+
+  constructor(private http:HttpClient) {}
+
+    find(){
+      console.log("Hola inventario");
+      return this.http.get('http://localhost:3000/api/products');
+    }
+
+    insertOne(obj){
+      return this.http.post('http://localhost:3000/api/products', obj);
+    }
+
+    updateOne(obj){
+      return this.http.put('http://localhost:3000/api/products', obj);
+    }
+
+    deleteOne(id){
+      return this.http.delete('http://localhost:3000/api/products/' + id);
+    }
 
 }
